#KDSK-109 refactoring LogoutModal
